Revert "Replace the session list with a tree view"
This reverts commit 19e36c7ba8569d7e3ce709ffb3cfa4d0ae5859cc.

It didn't put "Default Settings" at the top,
and I couldn't figure out how to make it do it.
The original author didn't seem interested in doing it...
https://github.com/leeor/putty/commit/43a24203de6c4f20693b60d7089408fdfbfc4b3f#commitcomment-1475756

I'm not even sure why it has to have a "Default Settings" entry..
And treating it as a special case in the code everywhere is stupid.
It should hide that entry, and make a button to "Load Defaults" / "Save Defaults" / "Reset Defaults", or something.